Standardized bucket and cooldown feature via interface/base class.
